client/requirements.txt
Kivy==1.11.1
PyYAML==5.1.2
